requirements.txt
google-genai>=1.0.0
SpeechRecognition>=3.10.0
sounddevice>=0.5.0
numpy>=1.24.0
keyboard>=0.13.5
pyperclip>=1.8.2
keyring>=25.0.0
pystray>=0.19.0
Pillow>=10.0.0